Prepared for ChampionDetailsFragment to Activity
