+++ app/app.iml
@@ -85,7 +85,6 @@
     <orderEntry type="library" exported="" name="support-v4-21.0.0" level="project" />
     <orderEntry type="library" exported="" name="picasso-2.3.4" level="project" />
     <orderEntry type="library" exported="" name="circularimageview-1.1" level="project" />
-    <orderEntry type="library" exported="" name="hlistview-1.2.2" level="project" />
     <orderEntry type="library" exported="" name="parallaxscroll-1.0" level="project" />
     <orderEntry type="library" exported="" name="okhttp-2.0.0" level="project" />
     <orderEntry type="library" exported="" name="okio-1.0.0" level="project" />
