+++ app/src/main/java/eu/mok/mokeulol/activities/ChampionDetailsActivity.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2014.
+ *
+ * Visit https://github.com/donmahallem/andlol for more info!
+ *
+ * Romanes eunt domus - Brian!
+ */
+
+package eu.mok.mokeulol.activities;
+
+/**
+ * Created by Don on 18.10.2014.
+ */
+public class ChampionDetailsActivity extends LeagueActivity {
+}
